+++ src/event-listeners/ChangeEventListener.ts
@@ -1,5 +1,5 @@
 import { createTargetedUserAction } from '../user-action/createTargetedUserAction'
-import { HTMLInputWithValue, TargetedUserAction, UserActionType } from '../types'
+import { HTMLInputWithValue, UserActionType } from '../types'
 import { sanitizeHTMLElementValue } from '../utils/sanitizeHTMLElementValue'
 import TargetedEventListener, { TargetEventListenerOptions } from './TargetedEventListener'
 import { isTextField } from '../utils/listeners'
@@ -15,7 +15,7 @@ class ChangeEventListener extends TargetedEventListener {
     if (isTextField(event.target)) return
 
     const elementTarget = event.target as HTMLInputWithValue
-    const userAction: TargetedUserAction | null = createTargetedUserAction(event, this.userActionType, this.options)
+    const userAction = createTargetedUserAction(event, this.userActionType, this.options)
     if (userAction != null) {
       userAction.target.value = sanitizeHTMLElementValue(elementTarget)
       this.userActionHandler.handle(userAction)