+++ src/event-listeners/ResizeEventListener.ts
@@ -1,6 +1,6 @@
-import TargetedEventListener from './TargetedEventListener'
 import { UserActionType } from '../types'
+import RepeatedEventListener from './RepeatedEventListener'
 
-export default class ResizeEventListener extends TargetedEventListener {
+export default class ResizeEventListener extends RepeatedEventListener {
   userActionType = UserActionType.Resize
 }