+++ src/event-listeners/TouchMoveEventListener.ts
@@ -1,6 +1,6 @@
-import TargetedEventListener from './TargetedEventListener'
 import { UserActionType } from '../types'
+import RepeatedEventListener from './RepeatedEventListener'
 
-export default class TouchMoveEventListener extends TargetedEventListener {
+export default class TouchMoveEventListener extends RepeatedEventListener {
   userActionType = UserActionType.TouchMove
 }